# docs/swagger.yaml
definitions:
  ds.RequestShip:
    properties:
      comment:
        type: string
      completionDate:
        type: string
      containers20ftCount:
        type: integer
      containers40ftCount:
        type: integer
      creationDate:
        type: string
      formationDate:
        type: string
      loadingTime:
        type: number
      requestShipID:
        type: integer
      ships:
        items:
          $ref: '#/definitions/ds.ShipInRequest'
        type: array
      status:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/ds.User'
        description: автозаполнение пользователя в заявках
      userID:
        type: integer
    type: object
  ds.Ship:
    properties:
      capacity:
        type: number
      containers:
        type: integer
      cranes:
        type: integer
      description:
        type: string
      draft:
        type: number
      length:
        type: number
      name:
        type: string
      photoURL:
        type: string
      shipID:
        type: integer
      width:
        type: number
    type: object
  ds.ShipInRequest:
    properties:
      requestShipID:
        type: integer
      ship:
        $ref: '#/definitions/ds.Ship'
      shipID:
        type: integer
      shipsCount:
        type: integer
    type: object
  ds.User:
    properties:
      cargoWeight:
        type: number
      contacts:
        type: string
      containers20ftCount:
        type: integer
      containers40ftCount:
        type: integer
      fio:
        type: string
      login:
        type: string
      password:
        type: string
      role:
        description: '"guest" | "creator" | "port_operator"'
        type: string
      userID:
        type: integer
    type: object
info:
  contact: {}
paths:
  /api/request_ship:
    get:
      description: Retrieve a list of requests with optional filters
      parameters:
      - description: Start date filter
        in: query
        name: start_date
        type: string
      - description: End date filter
        in: query
        name: end_date
        type: string
      - description: Status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ds.RequestShip'
            type: array
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Получить список заявок на расчет времени погрузки
      tags:
      - request_ships
  /api/request_ship/{id}:
    delete:
      description: Remove an entire request from the system
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string'
          schema:
            type: object
        "400":
          description: 'message: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Удаление всей заявки
      tags:
      - request_ships
    get:
      description: Retrieve details of a specific request with its ships
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'request_ship_id: int, status: string, creation_date: string,
            containers_20ft_count: int, containers_40ft_count: int, comment: string,
            loading_time: int, ships: []object'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "404":
          description: 'error: string'
          schema:
            type: object
      summary: Одна заявка на расчет времени погрузки
      tags:
      - request_ships
    put:
      consumes:
      - application/json
      description: Update fields of an existing request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Request updates
        in: body
        name: request
        required: true
        schema:
          properties:
            comment:
              type: string
            containers_20ft_count:
              type: integer
            containers_40ft_count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Изменение полей заявки
      tags:
      - request_ships
  /api/request_ship/{id}/completion:
    post:
      description: Allow port_operator to complete or reject a formed request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Action (complete or reject)
        in: formData
        name: action
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string, loading_time: int (if completed)'
          schema:
            type: object
        "400":
          description: 'description: string'
          schema:
            type: object
        "404":
          description: 'description: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Завершить или отклонить заявку (модератор)
      tags:
      - request_ships
  /api/request_ship/{id}/formation:
    put:
      description: Finalize a draft request by the creator
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string'
          schema:
            type: object
        "400":
          description: 'description: string'
          schema:
            type: object
        "404":
          description: 'description: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Сформировать заявку на расчет времени погрузки
      tags:
      - request_ships
  /api/request_ship/{id}/ships/{ship_id}:
    delete:
      description: Remove a ship from a specific request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ship ID
        in: path
        name: ship_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'description: string'
          schema:
            type: object
        "400":
          description: 'status: string, description: string'
          schema:
            type: object
        "500":
          description: 'status: string, description: string'
          schema:
            type: object
      summary: Удаление корабля из заявки
      tags:
      - request_ships
    put:
      consumes:
      - application/json
      description: Update the number of ships in a specific request
      parameters:
      - description: Request ID
        in: path
        name: id
        required: true
        type: integer
      - description: Ship ID
        in: path
        name: ship_id
        required: true
        type: integer
      - description: Updated ship count
        in: body
        name: request
        required: true
        schema:
          properties:
            ships_count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 'status: string, message: string'
          schema:
            type: object
        "400":
          description: 'description: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Обновление количества кораблей в заявке
      tags:
      - request_ships
  /api/request_ship/basket:
    get:
      description: Retrieve the count of ships in the user's draft request
      produces:
      - application/json
      responses:
        "200":
          description: 'data: {request_ship_id: int, ships_count: int}'
          schema:
            type: object
        "401":
          description: 'message: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Получить корзину запросов
      tags:
      - request_ships
  /api/ships:
    get:
      description: Retrieve a list of ships with optional filters
      parameters:
      - description: Ship name filter
        in: query
        name: name
        type: string
      - description: Minimum capacity filter
        in: query
        name: capacity
        type: string
      - description: Active status filter
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 'data: []ds.Ship, count: int'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Получить список кораблей
      tags:
      - ships
    post:
      consumes:
      - application/json
      description: Add a new ship to the system
      parameters:
      - description: Ship data
        in: body
        name: ship
        required: true
        schema:
          $ref: '#/definitions/ds.Ship'
      produces:
      - application/json
      responses:
        "201":
          description: 'data: ds.Ship'
          schema:
            type: object
        "400":
          description: 'error: string'
          schema:
            type: object
        "500":
          description: 'error: string'
          schema:
            type: object
      summary: Создать корабль
      tags:
      - ships
...
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer {ваш_JWT_токен}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"